requirements.txt
# ============================================================================
# RBAC Python 项目依赖
# 使用 Supabase 官方 Python Client
# ============================================================================

# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Supabase 官方客户端
supabase>=2.0.0

# JWT
PyJWT==2.8.0

# 密码加密
bcrypt==4.1.2

# 工具库
python-dateutil==2.8.2

# 测试
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# 代码质量
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# 日志
loguru==0.7.2
